Split nydus-utils into three crates
Split the nydus utils crate into three crates:
- nydus-error: a low-level common error handling code
- nydus-utils: digest algorithm and metrics for rafs and storage
- nydus-app: app utilities for nydusd and nydus-image

With this change, compilation of rafs and storage becomes much faster.
