+++ Cargo.toml
@@ -27,18 +27,21 @@ nix = "0.17"
 anyhow = "1.0.35"
 base64 = { version = ">=0.12.0" }
 rust-fsm = "0.4.0"
-rafs = { path = "rafs", features = ["backend-registry", "backend-oss"] }
-nydus-utils = { path = "utils" }
-nydus-api = { path = "api" }
 vm-memory = { version = ">=0.2.0", optional = true }
 chrono = "0.4.19"
-storage = { path = "storage" }
 openssl = { version = "0.10.35", features = ["vendored"] }
 
 event-manager = { git = "https://github.com/rust-vmm/event-manager.git", tag = "v0.2.0" }
 fuse-rs = { git = "https://github.com/cloud-hypervisor/fuse-backend-rs.git", optional = true, rev = "cfd2cca" }
 vhost-rs = { git = "https://github.com/cloud-hypervisor/vhost.git", branch = "dragonball", package = "vhost", optional = true }
-vhost-user-backend = { git = "https://github.com/cloud-hypervisor/vhost-user-backend.git", package = "vhost_user_backend", optional = true }
+vhost-user-backend = { git = "https://github.com/cloud-hypervisor/vhost-user-backend.git", package = "vhost_user_backend", rev = "78757bfa", optional = true }
+
+nydus-api = { path = "api" }
+nydus-app = { path = "app" }
+nydus-error = { path = "error" }
+nydus-utils = { path = "utils" }
+rafs = { path = "rafs", features = ["backend-registry", "backend-oss"] }
+storage = { path = "storage" }
 
 [dev-dependencies]
 sendfd = "0.3.3"
@@ -55,4 +58,4 @@ virtiofs = [
 ]
 
 [workspace]
-members = ["utils", "rafs", "api", "storage"]
+members = ["api", "app", "error", "rafs", "storage", "utils"]
