+++ app/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "nydus-app"
+version = "0.1.0"
+authors = ["The Nydus Developers"]
+description = "Application framework and utilities for Nydus"
+repository = "https://github.com/dragonflyoss/image-service"
+edition = "2018"
+build = "../app/build.rs"
+
+[build-dependencies]
+built = { version = "=0.4.3", features = ["chrono", "git2"] }
+
+[dependencies]
+flexi_logger = { version = "0.17" }
+libc = "0.2"
+log = "0.4.8"
+nix = "0.17"
+serde = { version = ">=1.0.27", features = ["serde_derive"] }
+
+nydus-error = { path = "../error" }
