+++ error/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "nydus-error"
+version = "0.1.0"
+description = "Error handling utilities for Nydus"
+repository = "https://github.com/dragonflyoss/image-service"
+authors = ["The Nydus Developers"]
+edition = "2018"
+
+[dependencies]
+backtrace = "0.3"
+httpdate = "1.0"
+log = "0.4.8"
+serde = { version = ">=1.0.27", features = ["serde_derive", "rc"] }
+serde_json = ">=1.0.9"
